+++ conf/globals.go
@@ -34,21 +34,99 @@ const (
 	ErrorAWSAuthNeeded     = "must provide AWS profile name or key pair"
 )
 
-// OCI Constants
+// Google Cloud Constants
 const (
-	OciDefaultProfile = "DEFAULT"
+	GCBucketACLAuthenticatedRead = "authenticatedread"
+	GCBucketACLPrivate           = "private"
+	GCBucketACLPublicRead        = "publicread"
+	GCBucketACLPublicReadWrite   = "publicreadwrite"
+	GCBucketACLProjectPrivate    = "projectprivate"
 )
 
 const (
-	ErrorOCICompartmentNotSpecified = "OCI compartment will be tenancy root"
-	ErrorOCIAuthNotSpecified        = "OCI auth not specified"
-	ErrorOCIStorageNotSpecified     = "OCI storage tier is not valid"
+	GCObjectACLAuthenticatedRead = "authenticatedread"
+	GCObjectACLPrivate           = "private"
+	GCObjectACLPublicRead        = "publicread"
+	GCObjectACLProjectPrivate    = "projectprivate"
+	GCObjectACLBucketOwnerFull   = "bucketownerfullcontrol"
+	GCObjectACLBucketOwnerRead   = "bucketownerread"
 )
 
 const (
-	OracleStorageTierStandard         = "standard"
-	OracleStorageTierInfrequentAccess = "infrequentaccess" // the case is strange because of the
-	OracleStorageTierArchive          = "archive"
+	GCStorageStandard = "STANDARD"
+	GCStorageNearline = "NEARLINE"
+	GCStorageColdline = "COLDLINE"
+	GCStorageArchive  = "ARCHIVE"
+)
+
+const (
+	GCLocationTypeRegion = "region"
+	GCLocationTypeDual   = "dual-region"
+	GCLocationTypeMulti  = "multi-region"
+)
+
+const (
+	GCRegionNANE1    = "NORTHAMERICA-NORTHEAST1"
+	GCRegionNANE2    = "NORTHAMERICA-NORTHEAST2"
+	GCRegionNAS1     = "NORTHAMERICA-SOUTH1"
+	GCRegionUSC1     = "US-CENTRAL1"
+	GCRegionUSE1     = "US-EAST1"
+	GCRegionUSE4     = "US-EAST4"
+	GCRegionUSE5     = "US-EAST5"
+	GCRegionUSS1     = "US-SOUTH1"
+	GCRegionUSW1     = "US-WEST1"
+	GCRegionUSW2     = "US-WEST2"
+	GCRegionUSW3     = "US-WEST3"
+	GCRegionUSW4     = "US-WEST4"
+	GCRegionSAE1     = "SOUTHAMERICA-EAST1"
+	GCRegionSAW1     = "SOUTHAMERICA-WEST1"
+	GCRegionEUC2     = "EUROPE-CENTRAL2"
+	GCRegionEUN1     = "EUROPE-NORTH1"
+	GCRegionEUSW1    = "EUROPE-SOUTHWEST1"
+	GCRegionEUW1     = "EUROPE-WEST1"
+	GCRegionEUW2     = "EUROPE-WEST2"
+	GCRegionEUW3     = "EUROPE-WEST3"
+	GCRegionEUW4     = "EUROPE-WEST4"
+	GCRegionEUW6     = "EUROPE-WEST6"
+	GCRegionEUW8     = "EUROPE-WEST8"
+	GCRegionEUW9     = "EUROPE-WEST9"
+	GCRegionEUW12    = "EUROPE-WEST12"
+	GCRegionASIAE1   = "ASIA-EAST1"
+	GCRegionASIAE2   = "ASIA-EAST2"
+	GCRegionASIANE1  = "ASIA-NORTHEAST1"
+	GCRegionASIANE2  = "ASIA-NORTHEAST2"
+	GCRegionASIANE3  = "ASIA-NORTHEAST3"
+	GCRegionASIASE1  = "ASIA-SOUTHEAST1"
+	GCRegionASIAS1   = "ASIA-SOUTH1"
+	GCRegionASIAS2   = "ASIA-SOUTH2"
+	GCRegionASIASE2  = "ASIA-SOUTHEAST2"
+	GCRegionMEC1     = "MIDDLEEAST-CENTRAL1"
+	GCRegionMEC2     = "MIDDLEEAST-CENTRAL2"
+	GCRegionMEW1     = "MIDDLEEAST-WEST1"
+	GCRegionAUSSE1   = "AUSTRALIA-SOUTHEAST1"
+	GCRegionAUSSE2   = "AUSTRALIA-SOUTHEAST2"
+	GCRegionAFRICAS1 = "AFRICA-SOUTH1"
+)
+
+const (
+	GCDualAsia = "ASIA1"
+	GCDualEur4 = "EUR4"
+	GCDualEur5 = "EUR5"
+	GCDualEur7 = "EUR7"
+	GCDualEur8 = "EUR8"
+	GCDualNAM4 = "NAM4"
+)
+
+const (
+	GCMultiAsia = "ASIA"
+	GCMultiEU   = "EU"
+	GCMultiUS   = "US"
+)
+
+const (
+	InvalidGCBucketACL    = "invalid google cloud bucket acl"
+	InvalidGCObjectACL    = "invalid google cloud object acl"
+	InvalidGCLocationType = "invalid google cloud location type"
 )
 
 // Linode Constants
@@ -95,6 +173,23 @@ const (
 const (
 	LinodeInvalidRegion = "invalid Linode region provided"
 	LinodeAuthNeeded    = "Linode authentication not specified"
+)
+
+// OCI Constants
+const (
+	OciDefaultProfile = "DEFAULT"
+)
+
+const (
+	ErrorOCICompartmentNotSpecified = "OCI compartment will be tenancy root"
+	ErrorOCIAuthNotSpecified        = "OCI auth not specified"
+	ErrorOCIStorageNotSpecified     = "OCI storage tier is not valid"
+)
+
+const (
+	OracleStorageTierStandard         = "standard"
+	OracleStorageTierInfrequentAccess = "infrequentaccess" // the case is strange because of the
+	OracleStorageTierArchive          = "archive"
 )
 
 // Conf Errors