+++ s3packs/providers/gcloud/client.go
@@ -1,0 +1,48 @@
+package gcloud
+
+import (
+	"context"
+	"fmt"
+
+	"cloud.google.com/go/storage"
+	"google.golang.org/api/option"
+)
+
+type GoogleClient struct {
+	Ctx         context.Context
+	Storage     *storage.Client
+	Bucket      *storage.BucketHandle
+	BucketAttrs *storage.BucketAttrs
+	cfg         *googleCfg
+}
+
+type googleCfg struct {
+	adc string
+}
+
+func (client *GoogleClient) getClient() error {
+
+	client.Ctx = context.Background()
+
+	c, err := storage.NewClient(client.Ctx, option.WithCredentialsFile(client.cfg.adc))
+	if err != nil {
+		return fmt.Errorf("error configuring storage client: %s", err.Error())
+	}
+
+	client.Storage = c
+
+	return nil
+
+}
+
+func (client *GoogleClient) getBucket(name string) {
+
+	b := client.Storage.Bucket(name)
+	client.Bucket = b
+	client.BucketAttrs, _ = b.Attrs(client.Ctx)
+
+}
+
+func (client *GoogleClient) refreshBucket() {
+	client.getBucket(client.Bucket.BucketName())
+}